--- lib/consts.ts
@@ -1,10 +1,0 @@
-export const ENVIRONMENT = {
-  HOSTED_ZONE_ID: process.env.HOSTED_ZONE_ID,
-  ACM_CERTIFICATE_ID: process.env.ACM_CERTIFICATE_ID,
-  LOGIN_LINKS_PASSWORD: process.env.LOGIN_LINKS_PASSWORD,
-  SESSION_SIGNATURE_SECRET_1: process.env.SESSION_SIGNATURE_SECRET_1,
-  COMMITS_TOKEN: process.env.COMMITS_TOKEN,
-  MONGO_CONNECTION: 'NOT_SET_NEEDS_OTHER_PR',
-  DEPLOYMENT_ENVIRONMENT: process.env.DEPLOYMENT_ENVIRONMENT,
-  NODE_ENV: process.env.NODE_ENV,
-};
